requirements.txt
docx
pdfplumber
dotenv
fastapi
uvicorn
python-dotenv
python-docx
faiss-cpu
sentence-transformers
langchain
langchain-community
streamlit
python-multipart
fpdf

